allow dots as part of table name
This is specially useful when combined with schemas in postgres. For
example, in a multi-tenant environment, a schema could be used per
tenant to ensure data separation between tenants. However, enqueue would
happen into a shared table in a shared schema. This way, tenants share
the infrastructure (in this case workers instead of one worker process
per tenant) but still have their own data island.